Split LuceneDev4002 into its own analyzer class
LuceneDev4002 (StackTraceHelper-driven NoInlining requirement) is a
distinct rule from 4000/4001 (NoInlining-as-no-op detection): it has a
different trigger (invocation vs. method declaration) and no code fix.
Extract it into its own analyzer + test class. Shared logic for
recognising the [MethodImpl(MethodImplOptions.NoInlining)] attribute,
empty bodies, and interface/abstract methods moves into
NoInliningAttributeHelper.
